+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/OnlineAsyncTaskAccelByteQueryOfferById.cpp
@@ -73,7 +73,8 @@ void FOnlineAsyncTaskAccelByteQueryOfferById::HandleGetItemByIds(const TArray<FA
 		Offer->DynamicFields.Add(TEXT("IsConsumable"), Item.EntitlementType == EAccelByteEntitlementType::CONSUMABLE ? TEXT("true") : TEXT("false"));
 		Offer->DynamicFields.Add(TEXT("Category"), Item.CategoryPath);
 		Offer->DynamicFields.Add(TEXT("Name"), Item.Name);
-		Offer->DynamicFields.Add(TEXT("ItemType"), FString::Printf(TEXT("%d"), (int32)Item.ItemType));
+		Offer->DynamicFields.Add(TEXT("ItemType"), FAccelByteUtilities::GetUEnumValueAsString(Item.ItemType));
+		Offer->DynamicFields.Add(TEXT("Sku"), Item.Sku);
 		if (Item.ItemType == EAccelByteItemType::COINS)
 		{
 			Offer->DynamicFields.Add(TEXT("TargetCurrencyCode"), Item.TargetCurrencyCode);