+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/SessionV1/OnlineAsyncTaskAccelByteUpdateV1GameSettings.h
@@ -1,0 +1,39 @@
+// Copyright (c) 2022 AccelByte Inc. All Rights Reserved.
+// This is licensed software from AccelByte Inc, for limitations
+// and restrictions contact your company contract manager.
+
+#pragma once
+
+#include "AsyncTasks/OnlineAsyncTaskAccelByte.h"
+#include "OnlineSubsystemAccelByteTypes.h"
+#include "OnlineSessionInterfaceV1AccelByte.h"
+
+/**
+ * Async task to update dedicated or P2P game settings through the SessionBrowser APIs.
+ */
+class FOnlineAsyncTaskAccelByteUpdateV1GameSettings : public FOnlineAsyncTaskAccelByte
+{
+public:
+
+	FOnlineAsyncTaskAccelByteUpdateV1GameSettings(FOnlineSubsystemAccelByte* const InABInterface, FName InSessionName, FOnlineSessionSettings& InUpdatedSessionSettings, bool InBShouldRefreshOnlineData);
+
+	virtual void Initialize() override;
+	virtual void TriggerDelegates() override;
+
+protected:
+
+	virtual const FString GetTaskName() const override
+	{
+		return TEXT("FOnlineAsyncTaskAccelByteUpdateGameSession");
+	}
+
+private:
+
+	FName SessionName;
+
+	FString SessionId;
+
+	FOnlineSessionSettings& UpdatedSessionSettings;
+
+	bool bShouldRefreshOnlineData;
+};